backend/requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
automata-lib==8.4.0
cached_method==0.1.0
coloraide==4.0
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
distlib==0.3.9
Django==5.1.7
executing==2.1.0
filelock==3.18.0
frozendict==2.4.6
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.1
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
psycopg>=3.1.8
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
virtualenv==20.29.3
wcwidth==0.2.13
gunicorn==20.1.0
python-dotenv>=0.19.0
djangorestframework==3.15.2
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0